CH/REfactoring User Directory
