--- src/db/mongoose.js
@@ -1,9 +1,0 @@
-const mongoose = require('mongoose')
-
-
-// database connection
-mongoose.connect(process.env.MONGO_URL, {
-    useNewUrlParser: true,
-    useCreateIndex: true,
-    useFindAndModify: false
-})
