+++ src/routes/AdminSkill.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = new express.Router();
 const admin_controller = require('../controllers/AdminSkill');
-const { isAdmin } = require('../middleware/adminAuth');
-const isUser = require('../middleware/userAuth');
+const { isAdmin } = require('../middleware/AdminAuth');
+const isUser = require('../middleware/UserAuth');
 
 //add skills
 router.post('/admin/add-skill-level', isAdmin, admin_controller.addSkills);
