+++ src/routes/AdminUser.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = new express.Router();
 const admin_controller = require('../controllers/AdminUser');
-const { isAdmin } = require('../middleware/adminAuth');
+const { isAdmin } = require('../middleware/AdminAuth');
 
 //get trainee profile
 // router.get('/admin/view/profile', isUser, admin_controller.view_user_profile );
